Pipeline

+++ src/Control/PreProcessor.java
@@ -12,11 +12,11 @@ public class PreProcessor {
 
     private static Timer timer = new Timer();
 
-    public static boolean getInterleavedFastq(String forward, String reverse, String name, String path, String readGroup) throws IOException {
+    public static boolean getInterleavedFastq(String forward, String reverse, String name, String path, String seqName) throws IOException {
         timer.start();
         System.out.println("(" + new Date().toString() + ") Start getting interleaved");
 
-        Process fastq2sam = Picard.fastQ2Sam(forward, reverse, path + name + ".ubam", readGroup, name);
+        Process fastq2sam = Picard.fastQ2Sam(forward, reverse, path + name + ".ubam", seqName, name);
         if (!waitforProcess(fastq2sam, "fastq2sam")) return false;
 
         Process markIll = Picard.markIlluminaAdapters(path + name + ".ubam", path + name + "marked.bam", path + name + "metric.txt");
@@ -48,12 +48,13 @@ public class PreProcessor {
         Process sam2Bam = Samtools.sam2BamParallel(path + name + ".sam", path + name + ".bam");
         if (!waitforProcess(sam2Bam, "sam2bam")) return false;
         new File(path + name + ".sam").delete();
+        new File(path + name + ".log").delete();
 
         Process merge = Picard.mergeBamAlignment(bamUnmapped, path + name + ".bam", path + name + "_merged.bam", reference);
         if (!waitforProcess(merge, "Merge Bam Alignment")) return false;
         new File(path + name + ".bam").delete();
         new File(path + name + "_merged.bam").renameTo(new File(path + name + ".bam"));
-        new File(path + name + "_merged.bai").renameTo(new File(path + name + ".bai"));
+        new File(path + name + "_merged.bai").renameTo(new File(path + name + ".bam.bai"));
 
         return afterBwa(reference,path,name);
     }
@@ -62,7 +63,7 @@ public class PreProcessor {
                                                           String reference, String name, String nameA, String nameB, String path) throws IOException, InterruptedException {
         timer.start();
 
-        System.out.println("(" + new Date().toString() + ") Start getting Bam from interleaved");
+        System.out.println("(" + new Date().toString() + ") Start getting Bam from interleaveds");
         Parameter[] parameters = {new Parameter('p', "")};
 
         BWAMEMAligner bwaA = new BWAMEMAligner(fastqA, "", reference, path + nameA + ".sam", path + nameA + ".log", parameters);
@@ -104,8 +105,10 @@ public class PreProcessor {
         new File(path + nameB + ".bam").delete();
         new File(path + nameB + ".bai").delete();
 
-        Process merge = Samtools.merge(path + name + ".sorted.bam", path + nameA + ".sorted.bam", path + nameB + ".sorted.bam");
+        Process merge = Samtools.merge(path + name + ".bam", path + nameA + ".sorted.bam", path + nameB + ".sorted.bam");
         if (!waitforProcess(merge, "Merging aligned")) return false;
+        Process indexMerged = Samtools.index( path + name + ".bam");
+        if (!waitforProcess(indexMerged, "index sort merged")) return false;
 
         new File(path + nameA + ".sorted.bam").delete();
         new File(path + nameA + ".sorted.bam.bai").delete();
@@ -137,20 +140,21 @@ public class PreProcessor {
         Process index = Samtools.index( path + name + ".sorted.bam");
         if (!waitforProcess(index, "index sort")) return false;
         new File(path + name + ".bam").delete();
-        new File(path + name + ".bai").delete();
+        new File(path + name + ".bam.bai").delete();
 
         Process mark = Picard.markDuplicates(path + name + ".sorted.bam", path + name + ".sortedDeDup.bam", path + name + ".dups");
         if (!waitforProcess(mark, "mark")) return false;
         new File(path + name + ".sorted.bam").delete();
+        new File(path + name + ".sorted.bam.bai").delete();
 
-        Process recall = GATK.baseReacalibrator(reference, path + name + ".sortedDeDup.bam", path + name + "recall_data.table");
+        Process recall = GATK.baseReacalibrator(reference, path + name + ".sortedDeDup.bam", path + name + ".recall_data.table");
         if (!waitforProcess(recall, "recall")) return false;
 
-        Process applyBQSR = GATK.applyBQSR(reference, path + name + ".sortedDeDup.bam", path + name + ".bam", path +name+ "recall_data.table", path);
+        Process applyBQSR = GATK.applyBQSR(reference, path + name + ".sortedDeDup.bam", path + name + ".bam", path +name+ ".recall_data.table", path);
         if (!waitforProcess(applyBQSR, "Apply BQSR")) return false;
         new File(path + name + ".sortedDeDup.bam").delete();
 
-        System.out.println("done");
+        System.out.println("********************\ndone with getting BAM******************\n");
         timer.reset();
         return true;
 
@@ -160,7 +164,7 @@ public class PreProcessor {
         try {
             int status = process.waitFor();
             timer.stop();
-            if (status == 0) System.out.println(name + " terminado con exito: " + timer.report());
+            if (status == 0) System.out.println("*********************\n " + name + " terminado con exito: " + timer.report());
             else {
                 System.out.println("error en " + name);
                 return false;
